Rename webhook module as handler
